enhancement!(model): store discriminators as u16s (#1068)
Instead of storing discriminators as Strings we can store them as u16s.
Fields will continue to deserialize from either strings or integers and
will continue to serialize to strings to match the Discord API.

This will have some memory usage improvements in some cases and may be
easier for users to handle.